backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
psycopg2-binary
SQLAlchemy
alembic
python-multipart